add startup and add priceCalculatorOption in config

+++ src/Workshop.Api/Bll/Models/GoodModels.cs
@@ -4,4 +4,4 @@ public record GoodModels(
     int Lenght,
     int Width,
     int Height,
-    double Weight);
+    decimal Weight);
