+++ src/Workshop.Api/Responses/V3/CargoResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace Workshop.Api.Responses.V3;
 
 public record CargoResponse(
-    double Volume,
-    double Weight);
+    decimal Volume,
+    decimal Weight);
