+++ src/Workshop.Api/Responses/V3/GoodProperties.cs
@@ -4,4 +4,4 @@ public record GoodProperties(
     int Lenght,
     int Width,
     int Height,
-    double Weight);
+    decimal Weight);
